+++ crates/sncast/tests/e2e/mod.rs
@@ -8,3 +8,4 @@ mod multicall;
 mod script;
 mod show_config;
 mod tx_status;
+mod verify;